// types/shopify.ts
export type Money = {
    amount: string;
    currencyCode: string;
};

export type ProductOption = {
    id: string;
    name: string;
    values: string[];
};

export type Edge<T> = {
    node: T;
};  

export type Connection<T> = {
    edges: Array<Edge<T>>;
};

export type ProductVariant = {
    id: string;
    title: string;
    availableForSale: boolean;
    selectedOptions: {
      name: string;
      value: string;
    }[];
    price: Money;
  };
  
  export type Image = {
    url: string;
    altText: string;
    width: number;
    height: number;
  };
// ...
export type CartProduct = {
    id: string;
    handle: string;
    title: string;
    featuredImage: Image;
};

export type ShopifyCart = {
    id: string | undefined;
    checkoutUrl: string;
    cost: {
        subtotalAmount: Money;
        totalAmount: Money;
        totalTaxAmount: Money;
    };
    lines: Connection<CartItem>;
    totalQuantity: number;
};

export type CartItem = {
    id: string | undefined;
    quantity: number;
    cost: {
      totalAmount: Money;
    };
    merchandise: {
      id: string;
      title: string;
      selectedOptions: {
        name: string;
        value: string;
      }[];
      product: CartProduct;
    };
  };